Bit more refactoring, preface to dark theme
