+++ packages/core/ui/App.js
@@ -1,11 +1,5 @@
-import AppBar from '@material-ui/core/AppBar'
-import CircularProgress from '@material-ui/core/CircularProgress'
-import Icon from '@material-ui/core/Icon'
-import IconButton from '@material-ui/core/IconButton'
-import Slide from '@material-ui/core/Slide'
-import Toolbar from '@material-ui/core/Toolbar'
-import Typography from '@material-ui/core/Typography'
 import { makeStyles } from '@material-ui/core/styles'
+import useMediaQuery from '@material-ui/core/useMediaQuery'
 
 import { observer, PropTypes } from 'mobx-react'
 import { isAlive } from 'mobx-state-tree'
@@ -15,8 +9,8 @@ import { withContentRect } from 'react-measure'
 import ErrorBoundary from 'react-error-boundary'
 
 import { inDevelopment } from '../util'
+import DrawerWidget from './DrawerWidget'
 import DevTools from './DevTools'
-import Drawer from './Drawer'
 import ErrorSnackbar from './ErrorSnackbar'
 
 const useStyles = makeStyles(theme => ({
@@ -45,81 +39,11 @@ const useStyles = makeStyles(theme => ({
   components: {
     display: 'block',
   },
-  drawerCloseButton: {
-    float: 'right',
-  },
-  drawerToolbar: {
-    paddingLeft: theme.spacing(2),
-  },
-  drawerToolbarCloseButton: {
-    flexGrow: 1,
-  },
-  drawerLoading: {
-    margin: theme.spacing(2),
-  },
 }))
 
-const DrawerWidget = observer(props => {
-  const { session } = props
-  const { visibleDrawerWidget, pluginManager } = session
-  const {
-    LazyReactComponent,
-    HeadingComponent,
-    heading,
-  } = pluginManager.getDrawerWidgetType(visibleDrawerWidget.type)
-  const classes = useStyles()
-
-  return (
-    <Drawer session={session} open={Boolean(session.activeDrawerWidgets.size)}>
-      <Slide direction="left" in>
-        <div className={classes.defaultDrawer}>
-          <AppBar position="static" color="secondary">
-            <Toolbar
-              variant="dense"
-              disableGutters
-              className={classes.drawerToolbar}
-            >
-              <Typography variant="h6" color="inherit">
-                {HeadingComponent ? (
-                  <HeadingComponent model={visibleDrawerWidget} />
-                ) : (
-                  heading || undefined
-                )}
-              </Typography>
-              <div className={classes.drawerToolbarCloseButton} />
-              <IconButton
-                className={classes.drawerCloseButton}
-                color="inherit"
-                aria-label="Close"
-                onClick={() => session.hideDrawerWidget(visibleDrawerWidget)}
-              >
-                <Icon fontSize="small">close</Icon>
-              </IconButton>
-            </Toolbar>
-          </AppBar>
-          <React.Suspense
-            fallback={
-              <CircularProgress
-                disableShrink
-                className={classes.drawerLoading}
-              />
-            }
-          >
-            <LazyReactComponent model={visibleDrawerWidget} session={session} />
-          </React.Suspense>
-        </div>
-      </Slide>
-    </Drawer>
-  )
-})
-
-DrawerWidget.propTypes = {
-  session: PropTypes.observableObject.isRequired,
-}
 function App({ contentRect, measureRef, session }) {
   const classes = useStyles()
   const { pluginManager } = session
-
   useEffect(() => {
     if (contentRect.bounds.width) {
       if (isAlive(session)) {
