+++ packages/core/ui/theme.js
@@ -1,50 +1,55 @@
 import { createMuiTheme } from '@material-ui/core/styles'
+// import useMediaQuery from '@material-ui/core/useMediaQuery'
 
-export default createMuiTheme({
-  palette: {
-    // type: 'dark',
-    primary: { main: '#0D233F' },
-    secondary: { main: '#721E63' },
-  },
-  overrides: {
-    MuiIcon: {
-      colorSecondary: {
-        color: '#135560',
-      },
+export default function useTheme() {
+  // current themes are not that good but we should support dark theme
+  // let prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)')
+  return createMuiTheme({
+    palette: {
+      // type: prefersDarkMode ? 'dark' : 'light',
+      primary: { main: '#0D233F' },
+      secondary: { main: '#721E63' },
     },
-    MuiButton: {
-      textSecondary: {
-        color: '#135560',
-      },
-    },
-    MuiFab: {
-      secondary: {
-        backgroundColor: '#FFB11D',
-      },
-    },
-    MuiExpansionPanelSummary: {
-      root: {
-        background: '#135560',
-        '&$expanded': {
-          // overrides the subclass e.g. .MuiExpansionPanelSummary-root-311.MuiExpansionPanelSummary-expanded-312
-          minHeight: 0,
-          margin: 0,
-          color: '#FFFFFF',
+    overrides: {
+      MuiIcon: {
+        colorSecondary: {
+          color: '#135560',
         },
-        margin: 0,
-        minHeight: 0,
-        padding: '0px 8px',
       },
-      content: {
-        '&$expanded': {
+      MuiButton: {
+        textSecondary: {
+          color: '#135560',
+        },
+      },
+      MuiFab: {
+        secondary: {
+          backgroundColor: '#FFB11D',
+        },
+      },
+      MuiExpansionPanelSummary: {
+        root: {
+          background: '#135560',
+          '&$expanded': {
+            // overrides the subclass e.g. .MuiExpansionPanelSummary-root-311.MuiExpansionPanelSummary-expanded-312
+            minHeight: 0,
+            margin: 0,
+            color: '#FFFFFF',
+          },
+          margin: 0,
+          minHeight: 0,
+          padding: '0px 8px',
+        },
+        content: {
+          '&$expanded': {
+            margin: '8px 0px',
+          },
+          color: '#FFFFFF',
           margin: '8px 0px',
         },
-        color: '#FFFFFF',
-        margin: '8px 0px',
-      },
-      expanded: {
-        // empty block needed to keep small
+        expanded: {
+          // empty block needed to keep small
+        },
       },
     },
-  },
-})
+  })
+}
