+++ packages/jbrowse-web/src/JBrowse.js
@@ -1,6 +1,6 @@
 import { readConfObject } from '@gmod/jbrowse-core/configuration'
 import '@gmod/jbrowse-core/fonts/material-icons.css'
-import { App, theme } from '@gmod/jbrowse-core/ui'
+import { App, useTheme } from '@gmod/jbrowse-core/ui'
 import {
   toUrlSafeB64,
   fromUrlSafeB64,
@@ -211,7 +211,7 @@ export default observer(({ config, initialState }) => {
     DisplayComponent = <App session={root.session} />
 
   return (
-    <ThemeProvider theme={theme}>
+    <ThemeProvider theme={useTheme()}>
       <CssBaseline />
       {DisplayComponent}
     </ThemeProvider>
